Move store to app folder
