+++ panel/src/app/store/index.js
