+++ panel/src/app/store/modules/content.js
