+++ panel/src/app/store/modules/languages.js
